redo generate command into a test
